+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Changelog
 
 v1.20.0 ()
 * Update for TYPO3 13.4 LTS
+* BC: showMenu() in ToolBox is not static anymore
 
 v1.19.3 (??.0?.2025)
  * Migrate label translation